git-svn-id: https://scm.nearinfinity.com/svn/incubator/super-lucene/trunk@16449 5c733ba6-9949-11dd-800e-6d5850b5fe75

+++ src/com/nearinfinity/blur/BlurHit.java
@@ -1,9 +1,7 @@
 package com.nearinfinity.blur;
 
 public interface BlurHit {
-	
-	float getScore();
+	double getScore();
 	String getId();
 	String getReason();
-
 }
